My Fix #1169
I completely remade the solar panel, taking as a basis
ElectrotineGenerator. Now the solar panel is a full-fledged block.
Current panel settings:
There is no dependence on time. From 6 to 18 works in full force. I just
do not like it when there are such situations. I like both the Advanced
Solar Panels for IC2.
The energy store was added in size of 10.
It does not work during the rain.
Through the glass works. Through the glass panels MultiPart does not
work.
To other changes.
Removed the coefficient of resistance of wires and in powerunits.scala.
Added a limitation of 120 volts in powerunits.scala and voltage pulling
by modulus. Dont know what for, but it seems to me so better.
I deleted the recipe for the rocket. He broke my client, and still in
the IC2 there are the best missile packages.
It seems I broke the master-branch. Now something is wrong with the
branches. There is no main branch in MrTJP / ProjectRed. Sorry.

А теперь тоже самое только на родном языке.
Я полностью переделал солнечную панель, взяв за основу
ElectrotineGenerator. Теперь панель солнечных батарей является
полноценной единицей.
Текущие настройки панели:
Нет зависимости от времени. От 6 до 18 работает в полную силу. Мне
просто не нравится, когда возникают такие ситуации. Мне нравятся оба
Advanced Solar Panels для IC2.
Хранилище энергии было добавлено в размере 10.
Он не работает во время дождя.
Через стекло работает. Через стеклянные панели MultiPart не работает.
К другим изменениям.
Удален коэффициент сопротивления проводов и в powerunits.scala.
Добавлено ограничение на 120 вольт в powerunits.scala и напряжения
берётся по модулю. Не знаю, для чего, но мне это кажется лучше.
Убрал рецепт джетпака. Он у меня крашнул клиент и всё равно в IC2 есть
лучше джетпаки.
Кажется я сломал мастер-ветку. Теперь что-то не так с ветвями. В MrTJP /
ProjectRed нет главной ветви.